Move test to test_full.py
